implementaçao de criptografia com bcrypt e token middleware com JWT

+++ controllers/usuariosController.js
@@ -1,16 +1,29 @@
 const { Router } = require("express")
-
-const { Usuarios, Sequelize } = require("../models")
+const { Usuarios } = require("../models")
 const { Habilidades } = require("../models")
 const { usuarioSchema } = require("../schemas/usuarioSchema")
-const Op = require("sequelize").Op
+const {verificaJWT} = require("../middlewares/jwtMiddleware")
 
+const Op = require("sequelize").Op
+const mysql = require("mysql2")
+const bcrypt = require("bcrypt")
 const router = Router()
 
+/////////////////////////////////////////////////////////////
+
+//FOI NECESSÁRIO PARA USAR SQL DIRETO COM O BANCO
+var con = mysql.createConnection({
+    host: "localhost",
+    user: "root",
+    password: "root",
+    database: "wisedb_sequelize",
+})
+
 //GET ALL
-router.get("/", async (req, res) => {
+router.get("/", verificaJWT, async (req, res) => {
     try {
         const usuarios = await Usuarios.findAll({
+            attributes: { exclude: ["password"] },
             include: [
                 {
                     model: Habilidades,
@@ -26,10 +39,35 @@ router.get("/", async (req, res) => {
     }
 })
 
+//GET ALL NÃO ADMINISTRADORES:
+router.get("/usuarioscomuns", verificaJWT, async (req, res) => {
+    try {
+        const usuarios = await Usuarios.findAll({
+            where: {
+                isAdm: false,
+            },
+            attributes: { exclude: ["password"] },
+            include: [
+                {
+                    model: Habilidades,
+                    as: "habilidades",
+                    through: { attributes: [] },
+                },
+            ],
+        })
+
+        res.status(200).json(usuarios)
+    } catch (erro) {
+        res.status(400).json(erro)
+        console.log(erro)
+    }
+})
+
 //GET UM SÓ USUARIO
-router.get("/:id", async (req, res) => {
+router.get("/:id", verificaJWT, async (req, res) => {
     try {
         const usuario = await Usuarios.findByPk(req.params.id, {
+            attributes: { exclude: ["password"] },
             include: [
                 {
                     model: Habilidades,
@@ -56,15 +94,14 @@ router.post("/", async (req, res) => {
         if (usuarioValido.error) throw new Error(usuarioValido.error)
 
         //VERIFICA SE HABILIDADE EXISTE
-
         const verificaHabilidade = await Habilidades.findAll({
             where: { id: req.body.habilidades },
         })
-        if (verificaHabilidade.length != req.body.habilidades.length) {
+
+        if (verificaHabilidade.length != req.body.habilidades.length)
             throw new Error(
                 "Id de alguma habilidade é inexistente, favor verificar dados..."
             )
-        }
 
         //VERIFICA SE JÁ EXISTE CADASTRO DE LOGIN, CPF E EMAIL EXISTENTES NO BANCO DE DADOS:
         const verificaSeExisteNoSistema = await Usuarios.findOne({
@@ -76,30 +113,33 @@ router.post("/", async (req, res) => {
                 ],
             },
         })
-        console.log("EXISTE NO SISTEMA:")
-    
 
-        if(verificaSeExisteNoSistema && verificaSeExisteNoSistema.dataValues.cpf == data.cpf) throw new Error("CPF já cadastrado")
-        if(verificaSeExisteNoSistema && verificaSeExisteNoSistema.dataValues.login == data.login) throw new Error("Login já cadastrado")
-        if(verificaSeExisteNoSistema && verificaSeExisteNoSistema.dataValues.email == data.email) throw new Error("Email já cadastrado")
+        if (verificaSeExisteNoSistema?.dataValues.cpf == data.cpf)
+            throw new Error("CPF já cadastrado")
+        if (verificaSeExisteNoSistema?.dataValues.login == data.login)
+            throw new Error("Login já cadastrado")
+        if (verificaSeExisteNoSistema?.dataValues.email == data.email)
+            throw new Error("Email já cadastrado")
 
-        // const loginExiste = await Usuarios.findOne({
-        //     where: {login: data.login },
-        // })
-        // const cpfExiste = await Usuarios.findOne({
-        //     where: {cpf: data.cpf },
-        // })
-        // const emailExiste = await Usuarios.findOne({
-        //     where: {email: data.email },
-        // })
-        // if (loginExiste != null) throw new Error("Login já cadastrado")
-        // if (cpfExiste != null) throw new Error("CPF já cadastrado")
-        // if (emailExiste != null) throw new Error("Email já cadastrado")
-
-        //EFETIVA O CADASTRO APÓS PASSAR PELAS LINHAS ACIMA SEM ERRO:
+        // EFETIVA O CADASTRO APÓS PASSAR PELAS LINHAS ACIMA SEM ERRO:
         if (usuarioValido) {
-            const newUsuario = await Usuarios.create(data)
+            //CRIPTOGRAFA PASSWORD
+            const hashedPassword = await bcrypt.hash(data.password, 10)
+
+            // FINALMENTE CRIA O USUARIO APÓS AS VALIDAÇÕES
+            const newUsuario = await Usuarios.create({
+                nome: data.nome,
+                cpf: data.cpf,
+                login: data.login,
+                password: hashedPassword,
+                dataDeNascimento: data.dataDeNascimento,
+                resetPassword: data.resetPassword,
+                email: data.email,
+                isAdm: data.isAdm,
+            })
             newUsuario.setHabilidades(habilidades)
+            console.log("DATA")
+            console.log(data)
 
             res.status(200).json({
                 message: "Usuario cadastrado com sucesso:",
@@ -116,16 +156,29 @@ router.post("/", async (req, res) => {
 })
 
 //DELETE
-router.delete("/:id", async (req, res) => {
+router.delete("/:id", verificaJWT, async (req, res) => {
     try {
+        const sql = `DELETE FROM Usuarios_Habilidades where usuarioId = ${req.params.id} `
+        const deleteRelation = con.query(sql)
+
+        const usuarioExistente = await Usuarios.findOne({
+            where: { id: req.params.id },
+        })
+
+        if (req.params.id != usuarioExistente.dataValues.id) {
+            throw new Error("Usuario inexistente")
+        }
+
         const deletedUsuario = await Usuarios.destroy({
             where: {
                 id: req.params.id,
             },
+            force: true,
         })
+
         res.status(200).json({
-            message: "usuario deletado com sucesso",
-            deletedUsuario: req.body,
+            message: `usuario deletado com sucesso`,
+            usuarioDeletado: usuarioExistente.dataValues,
         })
     } catch (erro) {
         res.status(400).json({
@@ -136,7 +189,7 @@ router.delete("/:id", async (req, res) => {
 })
 
 //UPDATE
-router.put("/:id", async (req, res) => {
+router.put("/:id", verificaJWT, async (req, res) => {
     try {
         const { habilidades, ...data } = req.body
 
@@ -166,6 +219,7 @@ router.put("/:id", async (req, res) => {
     } catch (erro) {
         res.status(500).json({ message: erro })
         console.log(erro)
+        
     }
 })
